extracts story mock data; fleshes out tests
