+++ libs/core/ui/src/lib/nested-checkboxes/nested-checkboxes.component.stories.ts
@@ -4,97 +4,10 @@ import { moduleMetadata, Story, Meta } from '@storybook/angular';
 import { CheckboxStates, NestedCheckboxesComponent, TreeProvider } from './nested-checkboxes.component';
 import { CheckboxComponent } from '../checkbox/checkbox.component';
 import { TreeComponent } from '../tree/tree.component';
+import { ITEM, Item, NestedItemTreeProvider } from '../../../.storybook/nested-checkboxes.data';
 
-interface Item {
-  id: string;
-  parentId?: string;
-  children?: Item[];
-}
 
 type NestedCheckboxesArgs = NestedCheckboxesComponent<Item>;
-
-const ITEM = {
-  id: 'Africa',
-  children: [
-    {
-      id: 'Southern Africa',
-      parentId: 'Africa',
-      children: [
-        {
-          id: 'Swaziland',
-          parentId: 'Southern Africa',
-        },
-        {
-          id: 'Namibia',
-          parentId: 'Southern Africa',
-        },
-      ],
-    },
-    {
-      id: 'Central Africa',
-      parentId: 'Africa',
-    },
-    {
-      id: 'Northern Africa',
-      parentId: 'Africa',
-      children: [
-        {
-          id: 'Morocco',
-          parentId: 'Northern Africa',
-          children: [
-            {
-              id: 'Marrakesh',
-              parentId: 'Morocco',
-            },
-            {
-              id: 'Fes',
-              parentId: 'Morocco',
-            },
-          ],
-        },
-      ],
-    },
-  ],
-}
-
-class NestedItemTreeProvider implements TreeProvider<Item> {
-  private _itemsKeyedById: Record<string, Item> = {};
-
-  constructor(item: Item) {
-    // set itemsKeyedById recursively
-    const items = [item];
-    while (items.length) {
-      const currentItem = items.shift();
-      if (currentItem) {
-        const currentItemId = this.getId(currentItem);
-        const currentItemChildren = this.getChildren(currentItem);
-        this._itemsKeyedById[currentItemId] = currentItem;
-        if (currentItemChildren.length) {
-          currentItemChildren.forEach((child) => {
-            items.push(child);
-          });
-        }
-      }
-    }
-  }
-
-  getId(item: Item): string {
-    return item.id;
-  }
-
-  getParent(item: Item): Item | undefined {
-    const parentId = item.parentId;
-    if (parentId) {
-      return this._itemsKeyedById[parentId];
-    }
-    return undefined;
-  }
-
-  getChildren(item: Item): Item[] {
-    return item.children || [];
-  }
-}
-
 
 export default {
   title: 'NestedCheckboxesComponent',
@@ -106,7 +19,7 @@ export default {
     })
   ],
   argTypes: {
-    // onChange: { action: 'clicked' },
+    onClick: { action: 'clicked' },
   }
 } as Meta<NestedCheckboxesComponent<Item>>;
 
@@ -117,8 +30,12 @@ const Template: Story<NestedCheckboxesArgs> = (args: NestedCheckboxesArgs) => ({
       [item]="item"
       [treeProvider]="treeProvider"
       [ngModel]="states"
-      (ngModelChange)="onChange($event)"
+      (ngModelChange)="states = $event; onClick($event)"
     ></atocha-nested-checkboxes>
+
+    <br />
+
+    <p>states: {{ states | json }}</p>
   `,
 });
